Reorganized files to have a proper package structure
