--- src/__init__.py
@@ -1,1 +1,0 @@
-from .benchmark_mediation import *
